refactor(admin): extract Transcription section from AppFormEditor (#1781)
Continues the incremental split of AppFormEditor.jsx tracked in #1781,
alongside the other section-extraction PRs (#1942, #2044, #2046, #2048,
#2049, #2050). Extracts the Transcription card (enable toggle, model
picker, input-source checkboxes, streaming toggle, max duration) into
client/src/features/admin/components/app-form/TranscriptionSection.jsx.
AppFormEditor.jsx shrinks from 2902 to 2739 lines. Pure extraction, no
behavior change intended.
